Make generated url readonly
At the moment the affiliateLink functunality is open to abuse. Any string can create a new database entry by just visiting a unique afffiliate link.

https://attendize.test/e/{id}/{event-name}?ref={unique-data}

Note: this commit does not fix this! this just a todo.
The better way to implement the feature is to create a database entry after a button press on (https://attendize.test/event/1/customize#affiliates). This way we reduce spam