fix 'The maximum number of tickets you can register' validation message
Thanks @jrckmcsb for the report

It was displaying the total remaining tickets rather than the max per
booking

It could still be used to leak ticket availability information if
tickets are close to selling out when quantity_remaining is less than
max_per_person. I did not fix this because the error message would be
confusing to end users and it looks like this was intended as a feature.